BLAN-1287 - moved text to i18 for some notifications - updated imorts for pinia plugin instance - i18n constant creation moved to a separate directory with plugins
