+++ src/i18n/ua/push-notifications.js
@@ -1,0 +1,10 @@
+export default {
+  event_has_been_ended: {
+    title: 'Подія закінчилася',
+    text: 'Подія "{eventName}" була закінчена, ви можете залишити відгук про подію та її учасників',
+  },
+  invite_user_to_event: {
+    title: 'Запрошення на подію',
+    text: ''
+  }
+};
