+++ src/main.js
@@ -1,5 +1,4 @@
 import { createApp } from 'vue';
-import { createI18n } from 'vue-i18n';
 
 import { createDeviceDetector } from 'next-vue-device-detector';
 
@@ -15,13 +14,13 @@ import Maska from './workers/phone-maska-worker';
 
 import App from './App.vue';
 import router from './router';
-import pinia from './plugins/pinia';
+import { i18n } from './plugins/i18n.plugin';
+import { pinia } from './plugins/pinia.plugin';
 import pkg from '../package';
-import { UA_LANGUAGE } from './i18n/ua';
 
 import { hintsDirective } from './directives';
 
-import './guards/tech-works-page-guard'
+import './guards/tech-works-page-guard';
 import './workers/map-worker/map.init';
 
 import './assets/styles/main.scss';
@@ -34,13 +33,6 @@ import 'cropperjs/dist/cropper.css';
 import '@vueform/slider/themes/default.css';
 import '@mcistudio/vue-colorpicker/dist/style.css';
 
-export const i18n = createI18n({
-  locale: 'ua',
-  allowComposition: true,
-  messages: {
-    ua: UA_LANGUAGE,
-  },
-});
 createLoader();
 
 const app = createApp(App);
