+++ src/plugins/i18n.plugin.js
@@ -1,0 +1,10 @@
+import { createI18n } from 'vue-i18n';
+import { UA_LANGUAGE } from '../i18n/ua';
+
+export const i18n = createI18n({
+  locale: 'ua',
+  allowComposition: true,
+  messages: {
+    ua: UA_LANGUAGE,
+  },
+});
