--- src/plugins/pinia.js
@@ -1,5 +1,0 @@
-import { createPinia } from 'pinia';
-
-const pinia = createPinia();
-
-export default pinia;
