+++ src/plugins/pinia.plugin.js
@@ -1,0 +1,3 @@
+import { createPinia } from 'pinia';
+
+export const pinia = createPinia();
