+++ src/utils/logOut.js
@@ -1,7 +1,7 @@
 import { accessToken, refreshToken } from '../workers/token-worker';
 import { ROUTES } from '../router/router.const';
 import router from '../router';
-import pinia from '../plugins/pinia';
+import { pinia } from '../plugins/pinia.plugin';
 import { useUserDataStore } from '../stores/userData';
 import { useTokensStore } from '../stores/tokens';
 import { useHeaderHeightStore } from '../stores/headerHeight';
