+++ src/workers/web-socket-worker/message-types/notifications/invite-user-to-event.js
@@ -13,6 +13,7 @@ import {
 } from '../../message.action.types';
 import { WebSocketTypes } from '../../web.socket.types';
 import { API } from '../../../api-worker/api.worker';
+import { i18n } from '../../../../plugins/i18n.plugin';
 import dayjs from 'dayjs';
 
 @AuthWebSocketMessage()
@@ -64,6 +65,6 @@ export class InviteUserToEventMessage extends InitialMessage {
   }
 
   createTitle() {
-    return 'Запрошення на подію';
+    return i18n.global.t('push_notifications.invite_user_to_event.title');
   }
 }
